Refactor/application

+++ app/controllers/articles_controller.rb
@@ -1,8 +1,8 @@
 class ArticlesController < ApplicationController
-  before_action :authenticate_user!, except: [:index, :show]
+  before_action :authenticate_user!, only: [:new, :create, :edit, :update, :destroy]
 
   def index
-    @articles = Article.all
+    @articles = Article.includes(:user)
   end
 
   def new
@@ -26,6 +26,9 @@ class ArticlesController < ApplicationController
 
   def edit
     @article = Article.find(params[:id])
+    if @article.user.id != current_user.id
+      redirect_to articles_path
+    end
   end
 
   def update
@@ -40,7 +43,9 @@ class ArticlesController < ApplicationController
 
   def destroy
     article = Article.find(params[:id])
-    article.destroy
+    if article.user.id == current_user.id
+      article.destroy
+    end
     redirect_to root_path
   end
